# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanZone football fan group API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  favoriteTeam?: string | null;
  xp: number;
  isAdmin: boolean;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  groupName?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  username?: string;
  bio?: string;
  favoriteTeam?: string;
  avatarUrl?: string;
}

export interface FanGroup {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  badgeColor?: string | null;
  /** @nullable */
  badgeEmoji?: string | null;
  memberCount: number;
  totalXp: number;
  isMember?: boolean;
  createdAt: string;
}

export interface FanGroupInput {
  /** @minLength 1 */
  name: string;
  description: string;
  badgeColor?: string;
  badgeEmoji?: string;
}

export interface GroupMember {
  id: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  favoriteTeam?: string | null;
  xp: number;
  joinedAt: string;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  settled: 'settled',
} as const;

export interface Match {
  id: number;
  title: string;
  description: string;
  liveUrl: string;
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  scheduledAt?: string | null;
  status: MatchStatus;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  pokeCount: number;
  chatCount?: number;
  predictionCount?: number;
  createdAt: string;
}

export type MatchInputStatus = typeof MatchInputStatus[keyof typeof MatchInputStatus];


export const MatchInputStatus = {
  upcoming: 'upcoming',
  live: 'live',
  settled: 'settled',
} as const;

export interface MatchInput {
  /** @minLength 1 */
  title: string;
  description: string;
  liveUrl: string;
  homeTeam: string;
  awayTeam: string;
  scheduledAt?: string;
  status?: MatchInputStatus;
}

export type MatchUpdateStatus = typeof MatchUpdateStatus[keyof typeof MatchUpdateStatus];


export const MatchUpdateStatus = {
  upcoming: 'upcoming',
  live: 'live',
  settled: 'settled',
} as const;

export interface MatchUpdate {
  title?: string;
  description?: string;
  liveUrl?: string;
  status?: MatchUpdateStatus;
}

export interface MatchSettlement {
  homeScore: number;
  awayScore: number;
}

export interface SettlementResult {
  match: Match;
  winnersCount: number;
  xpAwarded: number;
}

export interface ChatMessage {
  id: number;
  matchId: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ForumPost {
  id: number;
  matchId: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  title: string;
  content: string;
  replyCount: number;
  createdAt: string;
}

export interface ForumPostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
}

export interface ForumReply {
  id: number;
  postId: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface ForumReplyInput {
  /** @minLength 1 */
  content: string;
}

export interface Poke {
  id: number;
  matchId: number;
  fromUserId: number;
  fromUsername?: string;
  toUserId: number;
  toUsername?: string;
  createdAt: string;
}

export interface PokeInput {
  toUserId: number;
}

export interface PokeSummary {
  matchId: number;
  totalPokes: number;
  recentPokes: Poke[];
}

export interface Prediction {
  id: number;
  matchId: number;
  userId: number;
  username: string;
  homeScore: number;
  awayScore: number;
  /** @nullable */
  isCorrect?: boolean | null;
  /** @nullable */
  xpAwarded?: number | null;
  createdAt: string;
}

export interface PredictionInput {
  homeScore: number;
  awayScore: number;
}

/**
 * @nullable
 */
export type MatchPollMyVote = typeof MatchPollMyVote[keyof typeof MatchPollMyVote] | null;


export const MatchPollMyVote = {
  home_win: 'home_win',
  draw: 'draw',
  away_win: 'away_win',
} as const;

export interface MatchPoll {
  matchId: number;
  homeWin: number;
  draw: number;
  awayWin: number;
  totalVotes: number;
  /** @nullable */
  myVote?: MatchPollMyVote;
}

export type MatchPollVoteInputOutcome = typeof MatchPollVoteInputOutcome[keyof typeof MatchPollVoteInputOutcome];


export const MatchPollVoteInputOutcome = {
  home_win: 'home_win',
  draw: 'draw',
  away_win: 'away_win',
} as const;

export interface MatchPollVoteInput {
  outcome: MatchPollVoteInputOutcome;
}

export type MatchPollSummaryStatus = typeof MatchPollSummaryStatus[keyof typeof MatchPollSummaryStatus];


export const MatchPollSummaryStatus = {
  upcoming: 'upcoming',
  live: 'live',
  settled: 'settled',
} as const;

export interface MatchPollSummary {
  matchId: number;
  homeTeam: string;
  awayTeam: string;
  title?: string;
  status: MatchPollSummaryStatus;
  /** @nullable */
  scheduledAt?: string | null;
  poll: MatchPoll;
}

export interface GroupLeaderboardEntry {
  rank: number;
  groupId: number;
  groupName: string;
  /** @nullable */
  badgeColor?: string | null;
  /** @nullable */
  badgeEmoji?: string | null;
  memberCount: number;
  totalXp: number;
}

export interface UserLeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  favoriteTeam?: string | null;
  xp: number;
  /** @nullable */
  groupName?: string | null;
}

export interface AdminStats {
  totalUsers: number;
  totalGroups: number;
  totalMatches: number;
  liveMatches: number;
  totalPredictions: number;
  totalXpAwarded: number;
}

export type ListGroupsParams = {
search?: string;
};

export type ListMatchesParams = {
status?: ListMatchesStatus;
};

export type ListMatchesStatus = typeof ListMatchesStatus[keyof typeof ListMatchesStatus];


export const ListMatchesStatus = {
  live: 'live',
  upcoming: 'upcoming',
  settled: 'settled',
} as const;