# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FanZone football fan group API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{matchId}/poll:
    get:
      operationId: getMatchPoll
      tags: [polls]
      summary: Get poll results for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poll results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchPoll"
    post:
      operationId: voteMatchPoll
      tags: [polls]
      summary: Submit a vote in a match poll
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchPollVoteInput"
      responses:
        "200":
          description: Updated poll results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchPoll"
        "400":
          description: Invalid input or already voted
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    MatchPoll:
      type: object
      required: [matchId, homeWin, draw, awayWin, totalVotes]
      properties:
        matchId:
          type: integer
        homeWin:
          type: integer
        draw:
          type: integer
        awayWin:
          type: integer
        totalVotes:
          type: integer
        myVote:
          type: ["string", "null"]
          enum: [home_win, draw, away_win, null]
    MatchPollVoteInput:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [home_win, draw, away_win]
...