# lib/api-zod/src/generated/types/matchPoll.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanZone football fan group API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchPollMyVote } from './matchPollMyVote';

export interface MatchPoll {
  matchId: number;
  homeWin: number;
  draw: number;
  awayWin: number;
  totalVotes: number;
  /** @nullable */
  myVote?: MatchPollMyVote;
}