# lib/api-zod/src/generated/types/matchPollVoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanZone football fan group API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchPollVoteInputOutcome } from './matchPollVoteInputOutcome';

export interface MatchPollVoteInput {
  outcome: MatchPollVoteInputOutcome;
}